+++ internal/serveruntime/boot_state.go
@@ -211,6 +211,7 @@ type bootState struct {
 	effectiveEC      cluster.ECConfig
 	stopApply        chan struct{}
 	rebalancer       *cluster.Rebalancer
+	evacuator        *cluster.DataGroupEvacuator
 	loadReporter     *cluster.LoadReporter
 	loadReporterStor *cluster.NodeStatsStore